last

+++ db/migrate/20200627023615_create_cars.rb
@@ -2,7 +2,7 @@ class CreateCars < ActiveRecord::Migration[6.0]
   def change
     create_table :cars do |t|
       t.string :name
-      t.integer :kilometer
+      t.float :kilometer
       t.string :mantenimiento
 
       t.timestamps
